requirements.txt
bitstring~=4.0.2
Pillow~=9.5.0
pip~=23.1.2
attrs~=23.1.0
wheel~=0.40.0
tornado~=6.3.1
setuptools~=67.7.2